# documentacion/OpenAPI/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9000/",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/venta/{id}": {
      "get": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Devuelve una venta",
        "description": "Servicio encargado de retornar una venta con id ingresado por parámetro {id}.",
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Actualiza una Venta",
        "description": "Servicio encargado de actualizar la venta con el ID ingresado por parámetro, retornando la venta actualizada.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Venta"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Elimina una Venta",
        "description": "Servicio encargado de eliminar la venta con el ID ingresada por parámetro.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/venta": {
      "get": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Devuelve el listado: Ventas",
        "description": "Servicio encargado de devolver el listado de ventas y sus clientes asociados.",
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Persiste una Venta",
        "description": "Servicio encargado de persistir una venta, retornando la venta con el id asignado.",
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Venta"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/venta/ventasPorDia": {
      "get": {
        "tags": [
          "Venta-controller"
        ],
        "summary": "Devuelve el reporte de ventas por día.",
        "description": "Servicio encargado de devolver el reporte de ventas por día.",
        "operationId": "generarReporteDeVentasPorDia",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReporteVentasPorDia"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          }
        }
      },
      "ItemVenta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "venta": {
            "$ref": "#/components/schemas/Venta"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          }
        }
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Venta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemVenta"
            }
          }
        }
      },
      "ReporteVentasPorDia": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "nombreProducto": {
            "type": "string"
          },
          "cantidadVentas": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}